requirements.txt
# =========================================================================
# MediMind AI — Phase 1 dependencies
# Python: 3.12 (use `py -3.12 -m venv venv` on Windows; tested versions
#   3.12.x. Avoid 3.14 — PyTorch CUDA wheels and several Phase 2-3 libs
#   don't have 3.14 wheels yet as of May 2026.)
# All packages are open source and free. Pinning major.minor for stability.
# =========================================================================

# --- Backend (FastAPI stack) ---
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.20      # File uploads
python-dotenv==1.0.1

# --- Database ---
sqlalchemy==2.0.36
# psycopg2-binary==2.9.10     # Uncomment when switching to Postgres/Supabase

# --- Machine Learning (Phase 1) ---
scikit-learn==1.6.0
pandas==2.2.3
numpy==2.2.0
joblib==1.4.2                  # Model serialization
xgboost==2.1.3                 # Used in heart-disease module (Phase 2)

# --- Logging ---
structlog==24.4.0

# --- Frontend ---
streamlit==1.41.1
requests==2.32.3
plotly==5.24.1                 # For risk visualizations

# --- Dev / Testing ---
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1                   # For testing FastAPI endpoints

# =========================================================================
# Below are pinned but NOT installed by default — uncomment per phase
# =========================================================================

# --- Phase 2: Prescription OCR ---
pytesseract==0.3.13
Pillow==11.0.0
opencv-python-headless==4.10.0.84
pypdfium2==4.30.0
# easyocr deferred to v0.2.0 — adds ~250MB PyTorch dependency

# --- Phase 3: Medical Imaging ---
# Install torch + torchvision separately first via the CPU index URL:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
torchxrayvision==1.3.4
pydicom==3.0.1
grad-cam==1.5.4
# Already installed in Phase 2: opencv-python-headless, Pillow

# --- Phase 4: Chat / RAG ---
sentence-transformers==3.3.1
chromadb==0.5.23
huggingface-hub==0.27.0

# --- Phase 5: Orchestrator + Rate limiting ---
slowapi==0.1.9
